+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 
 # Version 1.1.2+7 RC
 
-- (iOS) Fixed automatic detection of "System Theme" availability.
+ - (iOS) Fixed automatic detection of "System Theme" availability.
+ - (Android) BETA added shortcut
 
 # Version 1.1.1+6 02-10-2020